backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anthropic>=0.40.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
tavily-python>=0.5.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
croniter>=3.0.0
lxml>=5.0.0
